# Update ToRpc method to replace async with Task for perf improvement